Rename "Sprite Animator" -> SpriteAnimator
